+++ movie/src/App.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Footer from "./components/Footer";
-import Main from './components/main';
+import Main from './components/Main';
 import Header from "./components/Header";
 
 export const App: React.FC = () => {
