+++ movie/src/components/Main/index.tsx
